feat: attention category 파일 분류
